+++ controller/editorapicontroller.php
@@ -1,7 +1,7 @@
 <?php
 /**
  *
- * (c) Copyright Ascensio System SIA 2023
+ * (c) Copyright Ascensio System SIA 2024
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -341,6 +341,11 @@ class EditorApiController extends OCSController {
         $fileStorage = $file->getStorage();
         if ($fileStorage->instanceOfStorage("\OCA\Files_Sharing\SharedStorage") || !empty($shareToken)) {
             $shareId = empty($share) ? $fileStorage->getShareId() : $share->getId();
+
+            $params["editorConfig"]["coEditing"] = [
+                "mode" => "strict",
+                "change" => false
+            ];
 
             $extraPermissions = null;
             if ($this->extraPermissions !== null) {