+++ CHANGELOG.md
@@ -7,16 +7,19 @@
 - fixed error display in the mobile application
 - fixed opening file without download access
 - Nextcloud v25 is no longer supported
-- Arabic and Serbian empty file templates
+- offline viewer for share link
+- updatable list of supported formats
 - When creating a form from docx, pdf is created instead of oform
 
 ## Added
 - advanced access rights for talk files and share links
-- reference data from coediting
 - selecting a document to combine from the storage
+- reference data from coediting
 - opening a reference data source
 - changing a reference data source
-- updatable list of supported formats
+- setting for disable editors cron check
+- Arabic and Serbian empty file templates
+- list of users to protect ranges of cells
 
 ## 7.9.4
 ## Changed